Don't exit for certain bits of dist.sh

+++ dist.sh
@@ -1,7 +1,7 @@
 set -o errexit
 rm -r build
 xcodebuild -configuration Release -project Audioscrobbler.xcodeproj
-rm -rf /Applications/Audioscrobbler.app
-killall Audioscrobbler
+test -d /Applications/Audioscrobbler.app && rm -rf /Applications/Audioscrobbler.app
+killall Audioscrobbler || true
 mv build/Release/Audioscrobbler.app /Applications
 open /Applications/Audioscrobbler.app
